+++ packages/Ecotone/composer.json
@@ -36,9 +36,9 @@
     },
     "require": {
         "php": "^8.0",
-        "friendsofphp/proxy-manager-lts": "^1.0.12",
         "psr/log": "^2.0|^3.0",
-        "ramsey/uuid": "^4.0"
+        "ramsey/uuid": "^4.0",
+        "friendsofphp/proxy-manager-lts": "^1.0.12"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
