+++ packages/Ecotone/src/Messaging/Config/Container/Definition.php
@@ -76,28 +76,3 @@ class Definition
         return $this->methodCalls;
     }
 }
-
-/**
- * @internal
- */
-class MethodCall
-{
-    public function __construct(private string $methodName, private array $arguments)
-    {
-    }
-
-    public function getMethodName(): string
-    {
-        return $this->methodName;
-    }
-
-    public function getArguments(): array
-    {
-        return $this->arguments;
-    }
-
-    public function replaceArguments(array $arguments): void
-    {
-        $this->arguments = $arguments;
-    }
-}
